testassigment/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.0
Django==4.2.6
django-cors-headers==4.3.0
djangorestframework==3.14.0
idna==3.4
install==1.3.5
numpy==1.26.4
pandas==2.2.1
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2024.1
urllib3==2.0.6
